[fix]: Update shipping fee amount

+++ Zenkoi.BLL/DTOs/ShippingDTOs/ShippingFeeBreakdownDTO.cs
@@ -2,17 +2,17 @@ namespace Zenkoi.BLL.DTOs.ShippingDTOs
 {
     public class ShippingFeeBreakdownDTO
     {
-        public decimal BoxFee { get; set; }
+        public int BoxFee { get; set; }
         public int? ShippingBoxId { get; set; }
         public string? ShippingBoxName { get; set; }
 
         public List<BoxSelection>? Boxes { get; set; }
 
-        public decimal DistanceFee { get; set; }
+        public int DistanceFee { get; set; }
         public int? ShippingDistanceId { get; set; }
         public decimal? DistanceKm { get; set; }
 
-        public decimal TotalShippingFee { get; set; }
+        public int TotalShippingFee { get; set; }
 
         public string? Notes { get; set; }
     }
